// app-frontend/src/app/store/actions/user-action-types.enum.ts
export enum UserActionTypesEnum {
  getUsers = '[Users] getUsers',
  getUsersSuccess = '[Users] getUsers Success',
  getUsersFailure = '[Users] getUsers Failure',

  getUser = '[Users] getUser',
  getUserSuccess = '[Users] getUser Success',
  getUserFailure = '[Users] getUser Failure',

  addUser = '[Users] addUser',
  addUserSuccess = '[Users] addUser Success',
  addUserFailure = '[Users] addUser Failure',

  deleteUser = '[Users] deleteUser',
  deleteUserSuccess = '[Users] deleteUser Success',
  deleteUserFailure = '[Users] deleteUser Failure',

  editUser = '[Users] editUser',
  editUserSuccess = '[Users] editUser Success',
  editUserFailure = '[Users] editUser Failure',
}